// src/interfaces/OrderTransaction.ts
export interface iResultTiktokItem {
  orderId?: string
  orderStatus?: string
  orderSubstatus?: string
  cancelationReturnType?: string
  normalOrPreOrder?: string
  skuId?: string
  sellerSku?: number
  productName?: string
  variation?: string
  quantity?: number
  skuQuantityOfReturn?: number
  skuUnitOriginalPrice?: number
  skuSubtotalBeforeDiscount?: number
  skuPlatformDiscount?: number
  skuSellerDiscount?: number
  skuSubtotalAfterDiscount?: number
  shippingFeeAfterDiscount?: number
  originalShippingFee?: number
  shippingFeeSellerDiscount?: number
  shippingFeePlatformDiscount?: number
  taxes?: number
  smallOrderFee?: number
  orderAmount?: number
  orderRefundAmount?: number
  createdTime?: string
  paidTime?: string
  rtsTime?: string
  shippedTime?: string
  deliveredTime?: string
  cancelledTime?: StringifyOptions
  cancelBy?: string
  cancelReason?: string
  fulfillmentType?: string
  warehouseName?: string
  trackingId?: string
  deliveryOption?: string
  shippingProviderName?: string
  buyerMessage?: string
  buyerUsername?: string
  recipient?: string
  phone?: string
  zipcode?: string
  country?: string
  province?: string
  district?: string
  detailAddress?: string
  additionalAddressInformation?: string
  paymentMethod?: string
  weight?: number
  productCategory?: string
  packageId?: string
  sellerNote?: string
  checkedStatus?: string
  checkedMarkedBy?: string
  profitCenter?: string
}